Added loss file, simplify objective in order to obtain a running algorithm
